add rbac

+++ src/http/routes/organizations/delete-organization.ts
@@ -3,7 +3,11 @@ import type { ZodTypeProvider } from 'fastify-type-provider-zod'
 import { z } from 'zod'
 
 import { auth } from '@/http/middlewares/auth'
+import { organizationSchema } from '@/lib/auth'
 import { prisma } from '@/lib/prisma'
+import { getUserPermissions } from '@/utils/get-user-permissions'
+
+import { UnauthorizedError } from '../_errors/unauthorized-error'
 
 export async function deleteOrganization(app: FastifyInstance) {
   app
@@ -27,17 +31,21 @@ export async function deleteOrganization(app: FastifyInstance) {
       async (request, reply) => {
         const { id } = request.params
         const userId = await request.getCurrentUserId()
-        const { organization } = await request.getUserMembership(id)
+        const { membership, organization } = await request.getUserMembership(id)
+
+        const authOrganization = organizationSchema.parse(organization)
+
+        const { cannot } = getUserPermissions(userId, membership.role)
+
+        if (cannot('delete', authOrganization)) {
+          throw new UnauthorizedError(
+            `You're not allowed to delete this organization.`,
+          )
+        }
 
         await prisma.organization.delete({
           where: {
             id: organization.id,
-            ownerId: userId,
-            members: {
-              some: {
-                role: 'ADMIN',
-              },
-            },
           },
         })
